+++ src/main/java/org/ssutt/core/fetch/entities/HTMLCellEntity.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.ssutt.core.dm.entities;
+package org.ssutt.core.fetch.entities;
 
 import java.util.ArrayList;
 import java.util.List;