+++ CHANGELOG.md
@@ -12,8 +12,11 @@ Changelog
 * `TTDataManager`:
 	+ Now `deliverDataFetcherProvider()` requiries `TTDataFetcher` implementation as a parameter. This way we can specify DF for any university outside the TT Core library and create real polymorphism. 
 	+ The same with `deliverDBProvider()`. Now requires `TTSQLManager` implementation and `Queries` implmenetation. This also allows to use TT Core for not only SSU TT Platform and not only with H2DB.
+	+ `getTT` now returns `List<String[]>` to provide working with raw strings data across DM fully and delegate JSON-formatting to TT Platform classes as new return format of timetables is used. See api_refernce in TT Platform repository (link later) fore more information on Public API.
 * `TTSQLManager`:
-    + Added `setQueries` to provide `Queries` instance with their definition.  
+    + Added `setQueries` to provide `Queries` instance with their definition.
+* `org.ssutt.core.dm.entities` has `TableEntity` and this factory <b>no more</b> due to updates in `TTDataManager.getTT`
+* `org.ssutt.core.dm.entities` and `org.ssutt.core.dm.TableParser` moved to `org.ssutt.core.fetch.entities` package as they handle only HTML data manipulations.
 
 
 ###TT Core 1.0.35